backend/requirements.txt
Django==5.2.3
djangorestframework==3.16.0
django-cors-headers==4.3.1
black==25.1.0